+++ packages/noalph-parser/src/index.ts
@@ -1,3 +1,3 @@
-export { parse } from './parser'
-export { tokenize } from './tokenizer'
-export type { ParseResult, Token, TokenType } from './types'
+export { parse } from './parser.js'
+export { tokenize } from './tokenizer.js'
+export type { ParseResult, Token, TokenType } from './types.js'